# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StillGood App API
  description: |
    API для приложения StillGood.

    ## Роли
    - **USER** — поиск, бронирование, заказы
    - **PARTNER** — управление боксами, аналитика
    - **ADMIN** — веб-панель модерации

    ## Аутентификация
    - Access Token: 30 минут
    - Refresh Token: 14 дней
    - JWT Bearer token в заголовке Authorization
  version: 1.0.0
  contact:
    name: StillGood Development Team
    email: team@stillgood.tech
  x-team:
    - role: Tech Lead, Android Developer
      email: dmitry@stillgood.tech
    - role: Backend Developer
      email: timofey@stillgood.tech
    - role: Backend Developer
      email: artem@stillgood.tech
servers:
  - url: https://api.stillgood.tech/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development
...
paths:
...
  /auth/register/partner:
    post:
      tags: [Auth]
      summary: Регистрация PARTNER (заявка)
      operationId: registerPartner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerRegisterRequest'
      responses:
        '201':
          description: Заявка подана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Невалидный email, слабый пароль, отсутствует deviceToken или тело запроса некорректно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка регистрации партнёра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/oauth:
    post:
      tags: [Auth]
      summary: OAuth вход (Google/Yandex) — только USER
      operationId: oauthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthRequest'
      responses:
        '200':
          description: Существующий пользователь
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AuthResponse'
                  - type: object
                    properties:
                      isNewUser:
                        type: boolean
                        example: false
        '201':
          description: Новый пользователь
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AuthResponse'
                  - type: object
                    properties:
                      isNewUser:
                        type: boolean
                        example: true
        '400':
          description: Невалидный OAuth-провайдер, токен провайдера, deviceToken или тело запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Аккаунт заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email занят другим провайдером
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка OAuth-входа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновление токенов
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Новые токены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Невалидное тело запроса или отсутствует refreshToken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Невалидный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Аккаунт заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка обновления токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Смена пароля (авторизован)
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Пароль изменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Неверный текущий пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не передан или недействителен JWT access-токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email/request:
    post:
      tags: [Auth]
      summary: Запрос OTP для верификации email
      operationId: requestEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEmailVerificationRequest'
      responses:
        '200':
          description: Код верификации отправлен, если email существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestEmailVerificationResponse'
        '400':
          description: Невалидный email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Слишком много запросов (код уже был отправлен недавно)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Не удалось доставить письмо с кодом верификации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Неожиданная ошибка запроса верификации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-email/confirm:
    post:
      tags: [Auth]
      summary: Подтверждение email по OTP-коду
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email подтверждён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Невалидный email или код
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Запрос OTP для сброса пароля
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP отправлен (если email существует)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  expiresIn:
                    type: integer
                    description: Секунды до истечения кода
                    example: 600
        '400':
          description: Невалидный email или тело запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Не удалось доставить письмо с кодом сброса пароля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Неожиданная ошибка запроса сброса пароля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/verify-reset-code:
    post:
      tags: [Auth]
      summary: Проверка OTP-кода
      operationId: verifyResetCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        '200':
          description: Код верный
          content:
            application/json:
              schema:
                type: object
                properties:
                  resetToken:
                    type: string
        '400':
          description: Неверный код
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Установка нового пароля
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, newPassword]
              properties:
                resetToken:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Пароль обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Невалидный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== USERS ====================
...
  /admin/users/{id}/block:
    post:
      tags: [Admin]
      summary: Заблокировать пользователя или партнёра
      operationId: blockUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Пользователь заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Нельзя заблокировать ADMIN или удалённый аккаунт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав (только ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{id}/unblock:
    post:
      tags: [Admin]
      summary: Разблокировать пользователя или партнёра
      operationId: unblockUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Пользователь разблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Нельзя разблокировать ADMIN или удалённый аккаунт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав (только ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PartnerRegisterRequest:
      type: object
      required: [email, password, name, phone, companyName, inn, establishmentName, establishmentAddress, deviceToken]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          description: ФИО контактного лица
        phone:
          type: string
        companyName:
          type: string
          description: Название юр. лица
        inn:
          type: string
        establishmentName:
          type: string
          description: Публичное название заведения
        establishmentAddress:
          type: string
        description:
          type: string
        deviceToken:
          type: string
          description: FCM device token для push-уведомлений
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Человекочитаемое описание
...